refac(build): fluidattacks#14 locate routes
- change routes directory
