+++ server/routes/testcases/test001.js
@@ -1,15 +1,13 @@
-const {
-  exec
-} = require("child_process");
-var express = require('express');
-var router = express.Router();
+import { exec } from "child_process";
+import { Router } from 'express';
+var router = Router();
 
 router.get('/test001/:user', function (req, res) {
   var user = req.params['user'];
   exec("ls target/user_files/" + user + "/", (error, stdout, stderr) => {
 
     var user_files = [];
-    for (file of stdout.split('\n')) {
+    for (var file of stdout.split('\n')) {
       if (file) {
         user_files.push(file);
       }
@@ -22,4 +20,4 @@ router.get('/test001/:user', function (req, res) {
   });
 });
 
-module.exports = router;
+export default router;
