+++ server/routes/testcases/test002.js
@@ -1,18 +1,18 @@
 const {
-  execFile
+  spawn
 } = require("child_process");
 var express = require('express');
 var router = express.Router();
 
-router.get('/test009/:user', function (req, res) {
+router.get('/test002/:user', function (req, res) {
   var user = req.params['user'];
 
-  var command = "/run/current-system/sw/bin/ls";
-  var arguments = ["target/user_files/" + user + "/"]
-  execFile(command, arguments, (error, stdout) => {
+  const ls = spawn("ls", ["target/user_files/" + user + "/"]);
 
+  ls.stdout.on("data", data => {
     var user_files = [];
-    for (file of stdout.split('\n')) {
+
+    for (var file of data.toString().split('\n')) {
       if (file) {
         user_files.push(file);
       }
@@ -22,6 +22,12 @@ router.get('/test009/:user', function (req, res) {
       'user_files': user_files
     };
     res.json(data);
+
+  });
+  ls.stderr.on("data", data => {
+    res.status(400).json({
+      'error': data.toString()
+    });
   });
 });
